fix: 1차 리팩
