+++ models/userDao.js
@@ -58,10 +58,45 @@ const findUserById = async userId => {
   return userInfo;
 };
 
+// 메일 변경
+const updateUserEmail = async (email, userId) => {
+  await myDataSource.query(
+    `
+
+    UPDATE
+      users
+    SET
+      email = '${email}'
+    WHERE
+      id = ${userId}
+
+
+    `
+  );
+};
+
+// 계정 탈퇴
+const deleteUserById = async (email, userId) => {
+  await myDataSource.query(
+    `
+
+    DELETE FROM
+        users
+    WHERE
+        email = '${email}'
+    AND
+        id = ${userId}
+
+    `
+  );
+};
+
 module.exports = {
   createUser,
   findUserByEmail,
   existUser,
   findUserByEmail,
   findUserById,
+  updateUserEmail,
+  deleteUserById,
 };
