+++ package.json
@@ -4,7 +4,8 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "nodemon server.js"
+    "start": "nodemon server.js",
+    "distribute": "npm i && npm run build && pm2 serve build 3000 --name justit --spa"
   },
   "author": "",
   "license": "ISC",
