+++ routes/scrapRouter.js
@@ -7,22 +7,10 @@ const {
   addPostScrap,
   findPostByUserId,
   deletePost,
-  updateUserEmail,
-  deleteUserById,
-} = require('../controllers/mypageController');
+} = require('../controllers/scrapController');
 
 router.post('/scrap', asyncWrap(mw.authMiddleware), asyncWrap(addPostScrap));
 router.get('/scrap', asyncWrap(mw.authMiddleware), asyncWrap(findPostByUserId));
 router.delete('/scrap', asyncWrap(mw.authMiddleware), asyncWrap(deletePost));
-router.put(
-  '/account',
-  asyncWrap(mw.authMiddleware),
-  asyncWrap(updateUserEmail)
-);
-router.delete(
-  '/leave',
-  asyncWrap(mw.authMiddleware),
-  asyncWrap(deleteUserById)
-);
 
 module.exports = router;
