+++ services/scrapService.js
@@ -1,4 +1,4 @@
-const mypageDao = require('../models/mypageDao');
+const mypageDao = require('../models/scrapDao');
 
 // 스크랩 추가
 const addPostScrap = async (posts_id, userId) => {
@@ -18,22 +18,8 @@ const deletePost = async (posts_id, userId) => {
   return deletePost;
 };
 
-// 메일 변경
-const updateUserEmail = async (email, userId) => {
-  const updateEmail = await mypageDao.updateUserEmail(email, userId);
-  return updateEmail;
-};
-
-// 계정 탈퇴
-const deleteUserById = async (email, userId) => {
-  const deleteUser = await mypageDao.deleteUserById(email, userId);
-  return deleteUser;
-};
-
 module.exports = {
   addPostScrap,
   findPostByUserId,
   deletePost,
-  updateUserEmail,
-  deleteUserById,
 };
