+++ services/userService.js
@@ -57,4 +57,16 @@ const getMe = async userId => {
   return userInfo;
 };
 
-module.exports = { signup, login, getMe };
+// 메일 변경
+const updateUserEmail = async (email, userId) => {
+  const updateEmail = await mypageDao.updateUserEmail(email, userId);
+  return updateEmail;
+};
+
+// 계정 탈퇴
+const deleteUserById = async (email, userId) => {
+  const deleteUser = await mypageDao.deleteUserById(email, userId);
+  return deleteUser;
+};
+
+module.exports = { signup, login, getMe, updateUserEmail, deleteUserById };
